Fixed the alternate rows so that both row types are being highlighted.
For Area Path, there could be 3 scenarios:
item path is under one of the paths in the config file
item path is under a sub node of one of the paths in the config file (this is the bug i fixed)
item path is under a node that is not part of the paths in the config file (i.e. )
for the 3rd case, we first need to add the 'unmapped' option to the combobbox (to be able to show it in the UI) and ignore if user select it. then if the user wants to map it to a different node, we allow it. Note: we assume that all nodes are under myConfig().RootNode.